monitoring/light_exporter/requirements.txt
prometheus-client==0.17.1
PyYAML==5.4.1
requests==2.31.0